+++ src/mesh3d/components.rs
@@ -25,8 +25,22 @@ impl Face {
     }
 
     /// Return a vector with the elements found at the vertex indices of the given slice
-    pub fn index_into<T: Copy>(&self, vertices: &[T]) -> Vec<T> {
-        // TODO: return `None` if the input slice isnt large enough
-        self.v_indices.iter().map(|vi| vertices[*vi]).collect()
+    ///
+    /// # Errors
+    /// Will return an error if the passed slice is not long enough to be indexed by the face
+    pub fn index_into<T: Copy>(&self, vertices: &[T]) -> Result<Vec<T>, String> {
+        let mut indexed_vertices = Vec::new();
+        for v_index in &self.v_indices {
+            if let Some(vertex) = vertices.get(*v_index) {
+                indexed_vertices.push(*vertex);
+            } else {
+                return Err(format!(
+                    "Mesh face vertex index ({}) is out of bounds ({})",
+                    v_index,
+                    vertices.len(),
+                ));
+            }
+        }
+        Ok(indexed_vertices)
     }
 }